+++ CHANGELOG.md
@@ -7,9 +7,9 @@
 - New Shortcut to toogle modal : Ctrl + /
 - When modal is opened, it will anounce page's snapshot info, continued about how to close the modal
 - When modal is closed, announce to user.
-- New shortcut to focus on the chat input box.
-- TODO: New shortcut to copy the last response (Ctrl + Shift + C)
-- TODO: New shortcut to announce the last response.
+- New shortcut to focus on the chat input box : `Shift + Esc`.
+- New shortcut to announce the last response : `Ctrl + Shift + Enter`.
+- TODO: New shortcut to copy the last response
 - TODO: Play sound as sign that a response is being received.
 - TODO: Detect if this is new chat. If not, get the title of current chat.
 - TODO: New shortcut to delete current chat / conversation.